Make text on iOS landing pages visible
The percentage padding on carouselTextContainer was causing problems
with display of the text on iOS. This has been made to be based on
Dimensions, as with the claims details page. Width is used for both
vertical and horizontal padding as using Dimensions height created
too large of a value and caused display errors with the text as well.