Position mnemonicPhrase center of screen
We are using position relative and absolute to put the mnemonic phrase
in the middle of the screen, using the length of the current word with
an average character length.
There may be a way to calculate the actual rendered length of the word
and use that to position more accurately.