Update claimDetails style
Removed the interface from claimDetailsStyles as it couples the styling
tightly and provides little benefit; I think it may have been a quick
fix for type issues with our Block/CenteredText components.

header style is composed within the file as it is not a pattern that
appears often across the codebase, in comparison to the buttons.

The buttons appear in this configuration multiple times across different
components and so are defined in a separate file.