Move all render logic to main render in consent
It was confusing having to look at all the class methods of the consent
component to figure out what was happening in each render method.
Hopefully having it all in one place makes it easier to make changes to the
UI. If it proves unwieldy, it may need more refactoring into separate
components.