Add sharedConstants, pass leftIcon in props
It made more sense to calculate the leftIcon once and pass it to each
component, so if an update needs to be made it can be made in one place.
The cards are all based on the same measurements so the sharedConstants
has been added to abstract some more out. This can probably be more
abstracted.