Update Landing image files
These JS files are where the properties of the landing images are defined.
A preset file has been made to provide the properties so that they can be
changed in one place instead of manually across each landing file. The
first image has also been updated to fit SE better.