Remove pagination counter from landing
The pagination counter required setting the state in onSnapToItem, which
caused a flickering effect on Android when going between the first and last
slides.

Described here:

https://github.com/jolocom/smartwallet-app/pull/1425#issuecomment-525240897

For now, the solution is to remove the pagination as it just loops around
and the landing slides are not that important. We may look into using an
alternative carousel or updating this in the future, however.