Fix oversized cards
The cardWrapper having `flex: 1` meant that it filled up any available
space. This was a problem with the issuerCard on the consent/share
credentials screen on high resolution screens. The consent screen has the
top section defined to take up 0.3 of the screen and the issuerCard would
take up as much of the space as it could.