Add debug mixin style
Previously we had a debug prop that would cause a debug style on Block
and other structure components. I have used a sort of debug style during
development of the SmartWallet and believe that it is useful. Having a
debug mixin that is already defined and that can just be imported is handy.